📦 NEW: Introduce Editor Queries and Mutations (#3)
* 👌 IMPROVE: Add authorId field for quick workaround
This is to solve the issue of "ConnectedRelationshipInconsistency", I call it CRI issue for a reason.

* 🐛 FIX: Cannot return null for non-nullable field Post.author

* 👌 IMPROVE: Make canUpdatePost bug free

* 👌 IMPROVE: Make posts field of user model, uncompulsory

* 📦 NEW: create updatePost mutation

* 👌 IMPROVE: Add slug field to Post model

* 📦 NEW: Write getPost query

* ❎ SAVE: save postsConnection query for pagination

* 📦 NEW: use context.db.mutation/query instead of context.prisma (#2)

* ❎ SAVE: Use Prisma from prisma-binding and not the generated one
The earlier structure caused some problems. I hope this won't. Also, need to change all queries and mutations to make things work.

* 👌 IMPROVE: Change all queries and mutation calls to db

* 📦 NEW: Stop using Prisma Client by prisma.
Use Prisma offered by Prisma Binding. This is just what wesbos uses. It fixes the connection bugs, unleash the power of graphQL!